more fixes for sanitizer

+++ include/gpac/tools.h
@@ -86,7 +86,7 @@ extern "C" {
  *	Macro formating a 4-character code (or 4CC) "abcd" as 0xAABBCCDD
 */
 #ifndef GF_4CC
-#define GF_4CC(a,b,c,d) (((a)<<24)|((b)<<16)|((c)<<8)|(d))
+#define GF_4CC(a,b,c,d) ((((u32)a)<<24)|(((u32)b)<<16)|(((u32)c)<<8)|((u32)d))
 #endif
 
 /*!
